requirements.txt
pycord
tinydb==4.4.0
topggpy==1.4.0
humanize==3.5.0
jishaku==1.20.0.220
python-dotenv==0.17.1
flask==2.0.1
